[Tests] Move delete & modify tests to async controller
